Move `waitUntilAll()` to `standards/event.ts`
